pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gCull"
version = "0.0.1"
description = "Tool to clean noisy Gaussians associated with sky in a 3DGS model"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License"
]

dependencies = [
    "jaxtyping>=0.3.2",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyYAML>=6.0.2",
    "rich>=14.0.0",
    "scikit-learn>=1.7.0",
    "setuptools>=78.1.1",
    "transformers>=4.53.1",
    "typing_extensions>=4.14.1"
]

[project.urls]
"Homepage" = "https://github.com/db-githb/gCull"

[project.scripts]
gcull = "gCullCLI.cli_main:main"

[tool.setuptools.packages.find]
include = ["gCullCLI*", "gCullCUDA*", "gCullMASK*", "gCullPY*"]
